chore: rename mcp/contracts.py -> mcp/types.py
Rinominato per eliminare la collisione di significato con
toolkit/contracts/ (path contract API pubblica).
mcp/types.py contiene TypedDict per le shape dei dati MCP.
